// frontend/types/anime-platform.model.d.ts
import { AccessType } from '../enums';
import { Anime } from './anime.model';
import { Platform } from './platform.model';

export type AnimePlatform = {
  id: number;
  animeId: number;
  platformId: number;
  linkId: number;
  accessType: AccessType;
  nextEpisodeAiringAt: string | null;
  lastEpisodeAiredAt: string | null;
  intervalInDays: number;
  episodeAired: number;
  isMainPlatform: boolean;
  isHiatus: boolean;
  platform: Platform;
  link: Link;
};

export type Link = {
  id: number;
  url: string;
  platformId: number;
};

export type AnimePlatformWithRelation = AnimePlatform & {
  anime: Anime;
};